// src/pages/sample-essay/ui/archive/types.ts
export interface SampleEssayBannerConfig {
  writing: {
    title: string;
    description: {
      line1: string;
      line2: string;
    };
    backgroundColor: string;
  };
  speaking: {
    title: string;
    description: {
      line1: string;
      line2: string;
    };
    backgroundColor: string;
  };
}